compatible with pytorch 0.5

+++ onmt/Trainer.py
@@ -42,10 +42,10 @@ class Statistics(object):
         self.n_correct += stat.n_correct
 
     def accuracy(self):
-        return 100 * (self.n_correct / self.n_words)
+        return 100 * (self.n_correct.to(dtype=torch.float) / self.n_words.to(dtype=torch.float))
 
     def ppl(self):
-        return math.exp(min(self.loss / self.n_words, 100))
+        return math.exp(min(self.loss / self.n_words.float(), 100))
 
     def elapsed_time(self):
         return time.time() - self.start_time
